eyetracking/src/requirements.txt
opencv-python
mediapipe
numpy
requests
pyyaml
python-velbus
google-genai
paho-mqtt
